Page Object package rename for correct structure
